moved pool ranking function to a new module
