+++ eras/shelley/test-suite/bench/Main.hs
@@ -40,7 +40,7 @@ import Cardano.Ledger.Shelley.LedgerState
     updateStakeDistribution,
   )
 import Cardano.Ledger.Shelley.PParams (PParams' (..))
-import Cardano.Ledger.Shelley.Rewards (likelihood)
+import Cardano.Ledger.Shelley.PoolRank (likelihood)
 import Cardano.Ledger.Shelley.UTxO (UTxO)
 import Cardano.Protocol.TPraos.API (PraosCrypto)
 import Cardano.Slotting.Slot (EpochSize (..))
